Conflicts:
	demos/pom.xml

+++ library/src/test/java/com/datatorrent/lib/algo/FilterValuesTest.java
@@ -1,11 +1,11 @@
-/*
- * Copyright (c) 2013 DataTorrent, Inc. ALL Rights Reserved.
+/**
+ * Copyright (C) 2015 DataTorrent, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ *         http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -18,7 +18,7 @@ package com.datatorrent.lib.algo;
 import java.util.ArrayList;
 import java.util.List;
 
-import junit.framework.Assert;
+import org.junit.Assert;
 
 import org.junit.Test;
 
@@ -37,7 +37,7 @@ public class FilterValuesTest
     ArrayList<Integer> ilist = (ArrayList<Integer>) list;
     int ret = 0;
     for (Integer i : ilist) {
-      ret += i.intValue();
+      ret += i;
     }
     return ret;
   }
