+++ library/src/test/java/com/datatorrent/lib/algo/UniqueCounterTest.java
@@ -1,11 +1,11 @@
-/*
- * Copyright (c) 2013 DataTorrent, Inc. ALL Rights Reserved.
+/**
+ * Copyright (C) 2015 DataTorrent, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ *         http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,7 +17,7 @@ package com.datatorrent.lib.algo;
 
 import java.util.HashMap;
 
-import junit.framework.Assert;
+import org.junit.Assert;
 
 import org.junit.Test;
 
@@ -66,11 +66,11 @@ public class UniqueCounterTest
     }
     oper.endWindow();
     HashMap<String,Integer> tuple = (HashMap<String,Integer>) sink.tuple;
-    int acount = tuple.get("a").intValue();
-    int bcount = tuple.get("b").intValue();
-    int ccount = tuple.get("c").intValue();
-    int dcount = tuple.get("d").intValue();
-    int ecount = tuple.get("e").intValue();
+    int acount = tuple.get("a");
+    int bcount = tuple.get("b");
+    int ccount = tuple.get("c");
+    int dcount = tuple.get("d");
+    int ecount = tuple.get("e");
     Assert.assertEquals("number emitted tuples", 1, sink.count);
     Assert.assertEquals("number emitted tuples", numTuples, acount);
     Assert.assertEquals("number emitted tuples", numTuples/2, bcount);
